requirements.txt
scikit-learn==1.4.2
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
biopython==1.85
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
pydantic==2.10.3
shap==0.45.1
matplotlib==3.8.4
Pillow==10.3.0